refactor(test): extract the duplicated Node 26 jsdom localStorage guard (#7620)
#7612's fix mirrored #7597's Node 26 jsdom/localStorage guard verbatim
into apps/loopover-ui and packages/loopover-ui-kit, on top of the
existing copy already in apps/loopover-miner-ui -- three copies of the
same ~15-line guard that can drift, and a fourth future jsdom
workspace wouldn't automatically inherit the fix.

Extract into test/helpers/vitest-jsdom-node26-localstorage.ts, exporting
one patchJsdomLocalStorageForNode26() function. All three vitest.setup.ts
files now import and call it instead of inlining the logic. No behavior
change -- same guard, same no-op-on-Node-22/24 / fix-on-Node-26+ semantics.

Verified on Node 26: all three consuming workspaces' full test suites
pass post-extraction (loopover-ui 79/79 files, ui-kit 3/3, ui-miner
28/28) -- the regression proof for a helper this thin, matching this
repo's existing test/helpers/*.ts convention of no dedicated test file.

Closes #7620
